chore: improve Divider, Drawer exports
